feat: Update Cloudflare zone resource to use `name` instead of `zone`
Upgrade Cloudflare provider to version 5.0

The changes in this commit update the Cloudflare zone resource to use the `name` attribute instead of the `zone` attribute, which is more accurate and consistent with the Cloudflare provider documentation. Additionally, the Terraform provider version for Cloudflare is upgraded to version 5.0 to ensure compatibility with the latest features and bug fixes.